do talked about util moving
