+++ Tasks/S3Download/package.json
@@ -3,9 +3,6 @@
     "version": "1.1.8",
     "description": "Download content from an AWS Simple Storage Service (S3) Bucket",
     "main": "S3Download.js",
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
     "author": "",
     "license": "MIT",
     "dependencies": {
