--- Tasks/S3Upload/tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ES6",
-        "module": "commonjs",
-        "sourceMap": true,
-        "alwaysStrict": true
-    },
-    "exclude": [
-        "node_modules"
-    ]
-}
