+++ tsconfig.json
@@ -9,8 +9,9 @@
         //"strict": true,
         "paths": {
             "aws-sdk*": [ "node_modules/aws-sdk/*" ],
-			"sdkutils*": [ "Tasks/Common/sdkutils/*" ],
-			"beanstalkutils*": [ "Tasks/Common/beanstalkutils*" ],
+            "Common*": [ "Tasks/Common*" ],
+            "sdkutils*": [ "Tasks/Common/sdkutils*" ],
+            "beanstalkutils*": [ "Tasks/Common/beanstalkutils*" ],
             "cloudformationutils*": [ "Tasks/Common/cloudformationutils*" ]
         },
         "rootDirs" : [
