+++ webpack.config.js
@@ -13,7 +13,8 @@ module.exports = {
 	alias: {
 	  "aws-sdk": path.resolve(__dirname, 'node_modules/aws-sdk'),
 	  "sdkutils": path.resolve(__dirname, '_build/Tasks/Common/sdkutils'),
-	  "beanstalkutils": path.resolve(__dirname, '_build/Tasks/Common/beanstalkutils'),
+    "Common": path.resolve(__dirname, '_build/Tasks/Common'),
+	  "beanstalkutils": path.resolve(__dirname, '_build/Tasks/Common/beanstalkutils')
 	}
   },
   externals: {
